Added labs 3,6,7,8

+++ Lab.5/Task_7/Task_7.cpp
@@ -30,7 +30,7 @@ int getValue()
 {
     while (true)
     {
-        int a;
+        unsigned long long int a;
         std::cin >> a;
         if (std::cin.fail())
         {
